+++ src/main/java/com/sparta/myblogserver/domain/post/repository/JdbcPostRepository.java
@@ -1,14 +1,13 @@
-package com.sparta.myblogserver.repository.post;
+package com.sparta.myblogserver.domain.post.repository;
 
-import com.sparta.myblogserver.entity.domain.dto.post.PostReq;
-import com.sparta.myblogserver.entity.domain.dto.post.PostRes;
-import com.sparta.myblogserver.entity.domain.entity.Post;
+import com.sparta.myblogserver.domain.post.dto.PostReq;
+import com.sparta.myblogserver.domain.post.dto.PostRes;
+import com.sparta.myblogserver.domain.post.entity.Post;
 import java.sql.Date;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
-import java.time.LocalDateTime;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.jdbc.core.JdbcTemplate;
